// src/app/business-logic/model/login/invite.ts
/**
 * login
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.12
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface Invite {
    /**
     * Invited email address
     */
    email?: string;
    /**
     * Time of invite
     */
    invited?: Date;
    /**
     * Invite status
     */
    status?: Invite.StatusEnum;
    /**
     * Time invite was accepted
     */
    accepted?: Date;
    user?: {
      id: string;
      role: string;
      name: string;
    };
}
export namespace Invite {
    export type StatusEnum = 'pending' | 'accepted' | 'timed_out';
    export const StatusEnum = {
        Pending: 'pending' as StatusEnum,
        Accepted: 'accepted' as StatusEnum,
        TimedOut: 'timed_out' as StatusEnum
    }
}