// src/app/business-logic/model/organization/organizationGetUserCompaniesResponseCompanies.ts
/**
 * organization
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.12
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { UserDetails } from './userDetails';


export interface OrganizationGetUserCompaniesResponseCompanies {
    /**
     * Company ID
     */
    id?: string;
    /**
     * Company name
     */
    name?: string;
    /**
     * Number of users allowed in company
     */
    allowed?: number;
    /**
     * Number of users allocated for company
     */
    allocated?: number;
    owners?: UserDetails[];
    /**
     * Company members (not including owners)
     */
    members?: Array<UserDetails>;
}